Add tracing functionality
Adds a `layer` function to the backend module which returns a
`tracing_subscriber::Layer` implementation configured to format events
in a format compatible with the log library used by the Go plugin
library, go-hclog. The returned `layer` can be installed as a
subscriber, potentially alongside other layers.

This uses a custom `tracing_subscriber::fmt::FormatEvent` implementation
which formats events created by `tracing` in a JSON with the correct
field names. Specifically this requires:

- timestamp in RFC3339 with microseconds and no 'Z', in a field named
  "@timestamp"
- level, in a field named "@level"
- message in a field named "@message"

Most of this commit is copying some code from the provided
`tracing_subscriber` JSON implementation with a few modifications.

We also instrument the various gRPC services with a debug trace, and
require that implementors also implement `std::fmt::Debug`.