Add `FrameValidity` marker trait and generic parameter to Frame
This uses the typestate pattern to ensure that a Frame is always in a
valid state before passing it back to the SDK, and provides suitable
methods on the relevant checked/unchecked Frame types.

Also adds a few convenience methods for creating frames more fluidly,
and plenty of examples to those methods.

An unfortunate casualty of this commit is the `MutableFrame` variant of
`StreamPacket`, which was incompatible with the way unchecking and
checking frames consumes the Frame.