Remove Frame type parameter; use new CheckedFrame<'a> instead
The marker type parameter in `Frame` was fun but a bit restrictive, in
that it required a new frame to be allocated on every iteration of a
stream. This commit removes it and introduces the `CheckedFrame<'a>`
struct instead, which is a simple wrapper around a reference to a `Frame`
that can only be constructed by `Frame::check`.

The various SDK structs that previously stored a `Frame` inside them
before serializing to JSON now do that serialization immediately, so
that they only ever require a reference to a `Frame` - or, now, a
`CheckedFrame`. This allows users to hold a mutable frame between
iterations while streaming, without requiring synchronization or `Arc`s,
at the expense of some additional error handling.