Bump prost to 0.11.0 and tonic to 0.8
This also removes the build dependencies on prost-build and tonic-build
because of the new requirement for `protoc` to be installed. Instead,
the generated code is committed as part of this PR, with instructions on
how to regenerate it in Cargo.toml. The build script is still present
but disabled using `build = false` in Cargo.toml.

This should speed up build times and reduces our dependencies, which is
great!