Pass owned request into DataService::query_data again
This is a bit of a tradeoff: users may need to clone aspects of the request
if they want to use it for every query in request.queries(), since the futures
they return need to outlive the request (and so must own the parts they need).
It does mean that users have ownership of the request again though... is that needed?