feat: add `GrafanaConfig` struct with a couple of useful keys
This adds a new struct, `GrafanaConfig`, to the existing `PluginContext`
struct.

The config can be used to obtain some useful fields provided by Grafana
such as the app URL and service account client secret, if available.

This commit doesn't include all of the possible keys that Grafana may
provide, but it does include the ones that are most likely to be useful
for me. If you need more, please open an issue or PR.